Extract Movies Detail components
